move Jsteam
